Keep in mind that Java isn't a dynamic language
The language that checks whether or not things have certain methods ON
THE FLY is JavaScript, not Java. Because of that you can't simply
combine a CurvatureDriveSystem and a MecanumDriveSystem in a ternary
expression and have the language which converts both objects to a
DriveSystem that doesn't have a public update() method figure out that
both objects define a public no-argument method called "update".